requirements.txt
google-generativeai==0.1.0
langchain==0.0.284
unstructured==0.10.12
pdf2image==1.16.3
pdfminer.six==20221105
streamlit==1.26.0
python-dotenv==1.0.0